Fix signedness of progress counter in incremental search example.
Switch the progress counter in incremental_search() to unsigned type,
matching the .len field, which it is compared against.  This is
possible because the cursor is always set at least to the beginning of
the current module id string, or NULL.  In the latter case,
incremental_search() is not even called.

Adjust all printf() format specifiers for unsigned variables.