Allow overriding what is passed as function context.
The __func__ macro, which is passed as the second argument to the
setup function, gets expanded to a string with the name of the calling
function.  In some applications, storing each possible context string
might be undesirable and needless overhead.

Introduce the macro DEBUG_MOD_CONTEXT to control what is passed as the
second argument.  If it is not defined before including debug_mod.h,
the default value is the __func__ magic macro.